[TASK] Protect TaskModuleController Part 2
TaskModuleController no longer extends BaseScriptClass. Properties
and methods are copied over to the controller and most of them are
set to protected. An instance of the controller class is given
to single class that fetch and set the one or the other information,
so not everything can be made protected. The patch should still
allow us to have more refactoring freedom in v10 in this area.

Resolves: #86204
Related: #86182
Releases: master
Change-Id: I171b22ddbcee67205234b6ca23acf9c95da6ac1b
Reviewed-on: https://review.typo3.org/58239
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>